+++ __tests__/StatisticsTest.js
@@ -1,5 +1,5 @@
 const { Console } = require("@woowacourse/mission-utils");
-const MatchingNumber = require("../src/MatchingNumber");
+const Statistics = require("../src/domain/Statistics");
 
 let totalLottoNumber = [
   [1, 5, 6, 10, 11, 13],
@@ -18,14 +18,14 @@ describe("로또 당첨 통계 부분 Test", () => {
   describe("5개 일치할때 보너스 번호 포함 테스트", () => {
     test("보너스번호가 포함되어 있지 않을경우 테스트", () => {
       const bonusNumber = "12";
-      const matchingNumber = new MatchingNumber(totalLottoNumber, winnerNumber, bonusNumber);
+      const matchingNumber = new Statistics(totalLottoNumber, winnerNumber, bonusNumber);
       const result = matchingNumber.isContainBonusNumber(numbers);
       expect(result).toBeFalsy();
     });
 
     test("보너스번호가 포함되어 있는 경우 테스트", () => {
       const bonusNumber = "13";
-      const matchingNumber = new MatchingNumber(totalLottoNumber, winnerNumber, bonusNumber);
+      const matchingNumber = new Statistics(totalLottoNumber, winnerNumber, bonusNumber);
       const result = matchingNumber.isContainBonusNumber(numbers);
       expect(result).toBeTruthy();
     });
@@ -38,28 +38,28 @@ describe("로또 당첨 통계 부분 Test", () => {
     });
     test("6개 일치하는 번호 테스트", () => {
       numbers = [1, 5, 6, 10, 11, 12];
-      const matchingNumber = new MatchingNumber(totalLottoNumber, winnerNumber, bonusNumber);
+      const matchingNumber = new Statistics(totalLottoNumber, winnerNumber, bonusNumber);
       const result = matchingNumber.getMatchingNumber(numbers);
       expect(result).toEqual([1, 5, 6, 10, 11, 12]);
     });
 
     test("5개 일치하는 번호 테스트", () => {
       let numbers = [1, 5, 6, 10, 11, 13];
-      const matchingNumber = new MatchingNumber(totalLottoNumber, winnerNumber, bonusNumber);
+      const matchingNumber = new Statistics(totalLottoNumber, winnerNumber, bonusNumber);
       const result = matchingNumber.getMatchingNumber(numbers);
       expect(result).toEqual([1, 5, 6, 10, 11]);
     });
 
     test("4개 일치하는 번호 테스트", () => {
       numbers = [1, 5, 6, 10, 15, 16];
-      const matchingNumber = new MatchingNumber(totalLottoNumber, winnerNumber, bonusNumber);
+      const matchingNumber = new Statistics(totalLottoNumber, winnerNumber, bonusNumber);
       const result = matchingNumber.getMatchingNumber(numbers);
       expect(result).toEqual([1, 5, 6, 10]);
     });
 
     test("3개 일치하는 번호 테스트", () => {
       numbers = [1, 5, 6, 8, 15, 16];
-      const matchingNumber = new MatchingNumber(totalLottoNumber, winnerNumber, bonusNumber);
+      const matchingNumber = new Statistics(totalLottoNumber, winnerNumber, bonusNumber);
       const result = matchingNumber.getMatchingNumber(numbers);
       expect(result).toEqual([1, 5, 6]);
     });
@@ -72,7 +72,7 @@ describe("로또 당첨 통계 부분 Test", () => {
         [1, 5, 6, 10, 11, 12],
       ];
       const bonusNumber = "13";
-      const matchingNumber = new MatchingNumber(totalLottoNumber, winnerNumber, bonusNumber);
+      const matchingNumber = new Statistics(totalLottoNumber, winnerNumber, bonusNumber);
       const result = matchingNumber.getResultOfThreeToFiveMatchingNumbers();
       expect(result).toEqual([0, 0, 0, 0, 2]);
     });
@@ -80,7 +80,7 @@ describe("로또 당첨 통계 부분 Test", () => {
     test("5개 일치 ,보너스 포함 총 3개", () => {
       const bonusNumber = "12";
       totalLottoNumber = [[1, 5, 6, 10, 12, 23], [1, 5, 6, 10, 12, 28], , [1, 5, 6, 10, 12, 45]];
-      const matchingNumber = new MatchingNumber(totalLottoNumber, winnerNumber, bonusNumber);
+      const matchingNumber = new Statistics(totalLottoNumber, winnerNumber, bonusNumber);
       const result = matchingNumber.getResultOfThreeToFiveMatchingNumbers();
       expect(result).toEqual([0, 0, 0, 3, 0]);
     });
@@ -88,7 +88,7 @@ describe("로또 당첨 통계 부분 Test", () => {
     test("5개 일치, 보너스 미포함 총 1개", () => {
       const bonusNumber = "32";
       totalLottoNumber = [[1, 5, 6, 10, 12, 23]];
-      const matchingNumber = new MatchingNumber(totalLottoNumber, winnerNumber, bonusNumber);
+      const matchingNumber = new Statistics(totalLottoNumber, winnerNumber, bonusNumber);
       const result = matchingNumber.getResultOfThreeToFiveMatchingNumbers();
       expect(result).toEqual([0, 0, 1, 0, 0]);
     });
@@ -100,7 +100,7 @@ describe("로또 당첨 통계 부분 Test", () => {
         [1, 5, 6, 18, 11, 42],
         [1, 5, 6, 12, 38, 43],
       ];
-      const matchingNumber = new MatchingNumber(totalLottoNumber, winnerNumber, bonusNumber);
+      const matchingNumber = new Statistics(totalLottoNumber, winnerNumber, bonusNumber);
       const result = matchingNumber.getResultOfThreeToFiveMatchingNumbers();
       expect(result).toEqual([0, 3, 0, 0, 0]);
     });
@@ -115,7 +115,7 @@ describe("로또 당첨 통계 부분 Test", () => {
         [1, 5, 12, 18, 26, 42],
         [1, 7, 8, 10, 11, 43],
       ];
-      const matchingNumber = new MatchingNumber(totalLottoNumber, winnerNumber, bonusNumber);
+      const matchingNumber = new Statistics(totalLottoNumber, winnerNumber, bonusNumber);
       const result = matchingNumber.getResultOfThreeToFiveMatchingNumbers();
       expect(result).toEqual([6, 0, 0, 0, 0]);
     });
